+++ assets/javascripts/kitten/components/buttons/button-image-with-text.js
@@ -25,7 +25,7 @@ class ButtonImageWithTextComponent extends Component {
       alt: '',
     },
     micro: false,
-    tag: '',
+    tag: 'p',
     largeGutter: false,
     regularTitle: false,
     regularText: false,